requirements.txt
MetaTrader5==5.0.45
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
tenacity==8.5.0
numpy==1.26.4
